Add unique constraint on unverified email addresses per user
In order to prevent users from DoS-ing others, a single unverified email
address is allowed.

This change put forth a quirk of the system: a user logged in who clicks
on the email verification link is redirected to the login page. Instead,
we could redirect them to their `edit_user_info` page.

The admin form validation used to depend on the order in which forms
were saved. `test_verify_email_ignores_other_emails` was working because
the `EmailAddress` for `me@mailinator.com` existed before the request,
and django saves new objects before existing objects.
The new validation ensures the unique constraints are met, and does not
rely on the order in which forms are processed.