Fix missing EmailInput.attrs.copy() in accounts forms
The `Widget.attrs` **must** be a copy, as the widget (or rendering
library) may mutate them.
https://github.com/django/django/blob/0ec60661e61b153e6bcec64649b1b7f524eb3e18/django/forms/widgets.py#L233

While potential bugs aren’t triggerable with the current version of the
codebase, it resulted in subtle bugs in upcoming work. A test would
cause an error in `RegisterForm`, and a following test would instantiate
another email field, which would have the `is-invalid` class.

Introduce a `verified_email_field()` to encourage reuse.