Fix verified email inline admin
Since the introduction of constraints (only one verified and unverified
email per user), the admin saving of email addresses breaks constraints
before to call `verify`, which would bring the situation back to a
normal state.

The main issue is using the `email` field as the PK of `EmailAddress`
prevents django from following that the instance did not change. When
foo@bar.com is changed to baz@bar.com, django cannot tell if baz@bar.com
is an existing instance being updated, or a new instance (with the old
one being delete). To help django keep track of EmailAddress instances,
introduce an `id` for EmailAddress.

Since the service is not live, rewrite the migration history instead of
crafting a real migration of the primary key, which is cumbersome.