Redirect to edit_user_info when CSRF failed
Reloading the page leads to a 404, when the user is on email confirm
view and has consumed the email session key in another tab/browser.
Instead of trying to be smart, juste direct users back to a valid page
in the application. If we had a landing page, that would be the ideal
target.