Add logging to Inclusion Connect
With the goal of sending the logs to ElasticSearch, logs are structured
with the JSON format.

Loggers are nested under the `inclusion_connect` root:
- `inclusion_connect.auth` handles authentication
- `inclusion_connect.oidc` handles the OpenID Connect protocol

Users are represented by their PK (UUID).
Applications are represented by their `client_id`.

Logs are sent `on_commit`, in an attempt to only log actions that
actually happened.

All logs have an `action` key, identifying what was performed. When the
acting user is different from the user (i.e. Django admin), an
`acting_user` key is added to the event.

When the client session is lost (changed browser), the OIDC application
is lost until the client resumes OIDC. The 'application' key is missing
from the logs in that case. That’s considered a minor inconvenience,
because:
- the application is logged when the user navigates on IC (until they
  lose their session, which often happens when they open the link from
  an email sent by IC)
- the user should immediately be redirected to their application when
  proceeding with OIDC authorize. The redirect is logged and allows deducing
  the relying party.

`caplog` was added to `oidc_flow_followup`, to ensure logging was
considered in all code paths.