Send logs to ElasticSearch
Logs are buffered by the application, and sent either:
- when the buffer is full, or
- after 5 seconds without logs

`flush()` does not send synchronously to ElasticSearch, to avoid
blocking on `emit()` (which calls `flush()` synchronously when capacity
is reached).

The `send_after_inactive_for_secs` is purposely not too large, to
increase the chances of flushing before the new logs come in. In the
worst case, with a capacity of 500, the logs can be retained 500 * 2s =
1000 seconds (a bit more than 15 minutes).

The Elasticsearch client implements a power of 2 backoff. With 10 max
retries, the app will wait for at 2^11-1 = 2047 seconds for
ElasticSearch to be back.