Use HTTP header to see when Scalingo is serving production
Because we need an email associated with connect.inclusion.beta.gouv.fr,
we cannot have a CNAME pointing to the production. AlwaysData only
supports an ALIAS for the apex of a domain that has an email.

Instead of a DNS solution, choose the next best thing: set a header when
serving from Scalingo, that the restore-backup script can read.