Convert existing IC @pe users to @ft emails
If we have an user using `@pole-emploi.fr` email, but without using the
PEAMA federation (not logged-in since the PEAMA federation was setup),
we may accidentally create a new user when PEAMA will return the new
email `@francetravail.fr`.

In this particular case, we consider that the user `user@pole-emploi.fr`
is the same than `user@francetravail.fr` and we update the old user with
the new email instead of creating a new user.

Co-authored-by: François Freitag <mail@franek.fr>