Set SESSION_COOKIE_SAMESITE="None"
Allow embedding the site in an iframe. Request from MonPortailPro, which
embeds multiple applications on a page. Because the session cookie was
SameSite=Lax, it would not be sent along with the authorize request when
users were directed to an application.

Setting SameSite=None is common for oauth providers, according to
https://auth0.com/docs/manage-users/cookies/samesite-cookie-attribute-changes

The Django CSRF middleware still protects from CSRF attacks, so the site
should remain protected.